+++ app/src/main/java/ru/netology/myapplication/dto/Step.kt
@@ -4,7 +4,7 @@ package ru.netology.myapplication.dto
 data class Step(
     val stepId: Long,
     val recipeIdStep: Long,
-    val stepOrder: Long,
+    val stepOrder: Int,
     val stepText: String,
     val stepImage: String? = null,
 )
